+++ libraries/matchers.rb
@@ -1,8 +1,7 @@
 #
 # Cookbook Name:: keepalived
-# Recipe:: default
 #
-# Copyright 2009, Chef Software, Inc.
+# Copyright 2016 Nathan Williams, <nath.e.will@gmail.com>
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -17,6 +16,18 @@
 # limitations under the License.
 #
 
-%w( install configure service ).each do |r|
-  include_recipe "#{cookbook_name}::#{r}"
+if defined?(ChefSpec)
+  require_relative 'keepalived'
+
+  Keepalived::RESOURCES.each do |r|
+    ChefSpec.define_matcher(r)
+
+    %w( create delete ).each do |a|
+      combo = "#{a}_#{r}".to_sym
+
+      define_method(combo) do |resource_name|
+        ChefSpec::Matchers::ResourceMatcher.new(combo, a.to_sym, resource_name)
+      end
+    end
+  end
 end