+++ spec/spec_helper.rb
@@ -1,4 +1,7 @@
 require 'chefspec'
 require 'chefspec/berkshelf'
 
+# Require all our libraries
+Dir.glob('libraries/*.rb').shuffle.each { |f| require File.expand_path(f) }
+
 at_exit { ChefSpec::Coverage.report! }